requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=10.0.0
scikit-learn>=1.3.0

# CLI and Interface
click>=8.0.0
rich>=13.0.0

# Data acquisition
kaggle>=1.5.0

# Type hints and development
mypy>=1.0.0
black>=23.0.0
ruff>=0.1.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Performance monitoring
tqdm>=4.65.0
psutil>=5.9.0